+++ tests/testthat/test_updateMeta.R
@@ -11,7 +11,6 @@ newDat <- df1$dat
 newDat$v3 <- c(4, 5)
 newDat$V1 <- NULL
 
-
 test_that("Remove rows meta helper", {
   expect_message(remove_rows_meta(df1$labels, names(newDat)), "Removing the following rows from meta data: V1")
   expect_equal(suppressMessages(remove_rows_meta(df1$labels, names(newDat))), df1$labels[df1$labels$varName == "ID1", ])
@@ -21,11 +20,13 @@ test_that("Remove rows meta helper", {
 })
 
 test_that("Add rows to meta helper", {
-  expect_message(add_rows_meta(df1$labels, newDat), "Adding meta data for the following variables: v3")
-  expect_equal(suppressMessages(add_rows_meta(df1$labels, newDat)), import_DF(newDat[, "v3", drop = F]))
+  expect_message(out <- add_rows_meta(df1$labels, newDat, checkVarNames = FALSE),
+                 "Adding meta data for the following variables: v3")
+  expect_equal(out, import_DF(newDat[, "v3", drop = F]))
 
-  expect_message(add_rows_meta(df1$labels, df1$dat), "No rows added to meta data.")
-  expect_equal(suppressMessages(add_rows_meta(df1$labels, df1$dat)), new_GADSdat(dat = data.frame(), labels = data.frame()))
+  expect_message(out2 <- add_rows_meta(df1$labels, df1$dat, checkVarNames = FALSE),
+                 "No rows added to meta data.")
+  expect_equal(out2, new_GADSdat(dat = data.frame(), labels = data.frame()))
 })
 
 
@@ -55,12 +56,14 @@ test_that("Update Meta all_GADSdat", {
   expect_equal(changes_out$allLabels$varName, c("ID1", "v3", "ID1", "V2", "v5", "v5"))
 })
 
-test_that("illegal variable names", {
+test_that("Invalid variable names are checked and changed", {
   newDat_ill <- df1$dat
   newDat_ill[, "Alter"] <- NA
   mess <- capture_messages(out_both <- updateMeta(df1, newDat_ill))
   expect_equal(mess[2], "Alter has been renamed to AlterVar\n")
   expect_equal(names(out_both$dat), c("ID1", "V1", "AlterVar"))
+  #out_both <- updateMeta(df1, newDat_ill)
+  #expect_equal(names(out_both$dat), c("ID1", "V1", "Alter"))
 })
 
 test_that("updateMeta extractData combination", {